backend/requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Data Science
pandas==2.1.4

# Machine Learning
scikit-learn==1.3.2

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0